move metric server to internal/app
